Bug 781530: don't check file hashes as part of cleanup r=bhearsum
It turns out that checking file hashes in cleanup() is a huge waste of time. We
already check that the file has the proper hash in the case that we need to
sign a previously uploaded file in response to a client request.

Having cleanup validate file hashes is therefore redundant, and puts a big io
load on the signing servers.