Bug 1170118 - Explicitly decode filenames in windows (clobberer); r=selenamarie
Mircosoft Multibyte Character Sets (MBCS) are not necessarily utf-8 compatible,
making it possible for clobberer to come across filenames which it doesn't know
how to read.

To handle this, the file system encoding scheme (MBCS) on our build machines is
detected, and used to explictly decode results from directory listings.

https://msdn.microsoft.com/en-us/library/ey142t48.aspx